se corrige la funcion de eliminar libro y se condiciona internamente

+++ src/main/java/com/ceiba/biblio/service/LibroService.java
@@ -22,15 +22,15 @@ import org.springframework.http.HttpHeaders;
 
 @Service
 public class LibroService {
-    
+
     @Autowired
     private LibroRepository libroRepository;
-    
+
     @Autowired
     private PrestamoRepository prestamoRepository;
-    
-    public ResponseEntity getInfo(){
-        
+
+    public ResponseEntity getInfo() {
+
         return ResponseEntity.ok("Respuesta servicio OK");
     }
 
@@ -39,12 +39,12 @@ public class LibroService {
         LibroOutDto respuesta = new LibroOutDto();
         Optional<LibroEntity> lib = libroRepository.findByIsbn(libro.getIsbn());
 
-        if(lib.isPresent()){
-            libr=lib.get();
-            libr.setTotalEjemplares(libr.getTotalEjemplares()+1);
-            libr.setTotalEjemplaresDisponibles(libr.getTotalEjemplaresDisponibles()+1);
+        if (lib.isPresent()) {
+            libr = lib.get();
+            libr.setTotalEjemplares(libr.getTotalEjemplares() + 1);
+            libr.setTotalEjemplaresDisponibles(libr.getTotalEjemplaresDisponibles() + 1);
             respuesta.setMensaje("Libro guardado correctamente");
-        }else {
+        } else {
             libr = new LibroEntity();
             libr.setTotalEjemplares(Long.valueOf(1));
             libr.setTotalEjemplaresDisponibles(Long.valueOf(1));
@@ -60,48 +60,62 @@ public class LibroService {
         return ResponseEntity.ok(respuesta);
 
     }
-    
+
     public ResponseEntity eliminarLibro(LibroInDto libro) {
         LibroEntity libr = null;
         LibroOutDto respuesta = new LibroOutDto();
         Optional<LibroEntity> lib = libroRepository.findByIsbn(libro.getIsbn());
 
-        if(lib.isPresent()){
-            libr=lib.get();
-            libr.setTotalEjemplares(libr.getTotalEjemplares()+1);
-            libr.setTotalEjemplaresDisponibles(libr.getTotalEjemplaresDisponibles()+1);
-            respuesta.setEstado("OK");
-            respuesta.setMensaje("Libro eliminado correctamente");
-            libroRepository.delete(libr);
-        }else {
+        if (lib.isPresent()) {
+            libr = lib.get();
+            if (libr.getTotalEjemplares() > 1) {
+                libr.setTotalEjemplares(libr.getTotalEjemplares() - 1);
+                libr.setTotalEjemplaresDisponibles(libr.getTotalEjemplaresDisponibles() - 1);
+                respuesta.setEstado("OK");
+                respuesta.setMensaje("Libro eliminado correctamente");
+                libroRepository.save(libr);
+            } else if (libr.getTotalEjemplaresDisponibles() == 0
+                    && libr.getTotalEjemplaresPrestados() == 0
+                    && libr.getTotalEjemplares() == 1) {
+                libr.setTotalEjemplares(libr.getTotalEjemplares() + 1);
+                libr.setTotalEjemplaresDisponibles(libr.getTotalEjemplaresDisponibles() + 1);
+                respuesta.setEstado("OK");
+                respuesta.setMensaje("Libro eliminado correctamente");
+                libroRepository.delete(libr);
+            }else if(libr.getTotalEjemplaresDisponibles()==0 && libr.getTotalEjemplaresPrestados()>0){
+                respuesta.setEstado("MAL");
+                respuesta.setMensaje("Libro no se puede eliminar. hay libros prestados.");
+            }
+
+        } else {
             respuesta.setEstado("MAL");
             respuesta.setMensaje("Libro no eliminado, no Existe");
         }
         return ResponseEntity.ok(respuesta);
     }
 
-    public boolean esPalindrome(String isbn){
-        String isbnTemporal  = isbn.replaceAll("\\s+", "").toLowerCase();
+    public boolean esPalindrome(String isbn) {
+        String isbnTemporal = isbn.replaceAll("\\s+", "").toLowerCase();
         return IntStream.range(0, isbnTemporal.length() / 2)
                 .noneMatch(i -> isbnTemporal.charAt(i) != isbnTemporal.charAt(isbnTemporal.length() - i - 1));
     }
-    
-    public ResponseEntity consultarLibros(){
-        
+
+    public ResponseEntity consultarLibros() {
+
         List<ListaLibroOutDto> listadoLibros = new ArrayList<>();
 
         List<LibroEntity> lib = libroRepository.findAll();
-        for(LibroEntity i : lib){
+        for (LibroEntity i : lib) {
             ListaLibroOutDto libDto = new ListaLibroOutDto();
             List<PrestamoEntity> prestamo = prestamoRepository.findByLibId(i);
-            if(prestamo.isEmpty()){
+            if (prestamo.isEmpty()) {
                 libDto.setFechaLimite(null);
                 libDto.setNombrePersona("");
                 libDto.setNombreLibro(i.getTitulo());
                 libDto.setIsbn(i.getIsbn());
                 listadoLibros.add(libDto);
-            }else{
-                for(PrestamoEntity p : prestamo){
+            } else {
+                for (PrestamoEntity p : prestamo) {
                     libDto = new ListaLibroOutDto();
                     libDto.setFechaLimite(p.getPresFechalimite());
                     libDto.setNombrePersona(p.getPresNombrepersona());
@@ -111,7 +125,7 @@ public class LibroService {
                 }
             }
         }
-        
+
         return ResponseEntity.ok(listadoLibros);
     }
 }
